Mark webhook and controller as safe-to-evict
This is a port of tektoncd/pipelines#4124

The safe-to-evict annotation tells the cluster autoscaler whether the
pod can be evicted to allow the node it's on to scale down.

Unfortunately, blocking node eviction means the node that the pod(s) get
scheduled to can't be scaled down. Furthermore, the nodes can't be fully
drained when updating the cluster. This can leave a cluster in a
mid-upgrade state that can make issues difficult to diagnose and reason
about.

With this change, a cluster scale-down event might cause temporary
service unreliability with the default single-replica configuration

Signed-off-by: Dibyo Mukherjee <dibyo@google.com>
(cherry picked from commit db351242647506e23695826cd8054ecfe981ea0b)
Signed-off-by: Vincent Demeester <vdemeest@redhat.com>